Add Task API endpoint to expose enrollments
Add an endpoint, `/Task/<id>`, where ID is the ID of an enrollment (also
known as ticket ID; the primary key of the `enrollments` table). Each
enrollment is represented as a FHIR Task[1], with the following fields
filled out:

- `identifier` - the enrollment application ID. Note that this is not
  the NPI, as a single provider may need to apply to enroll multiple
  times for different services.
- `status` - one of `Requested`, `Accepted`, `Rejected`
- `intent` (required by FHIR) - `proposal`
- `requester` - a reference to either a `Practitioner`[2] (individual
  provider) or an `Organization`[3] (group provider)
- `input` - represents the provider type; specifically, the `type` will
  be a `CodeableConcept`[4] with the `text` field set to "Provider
  Type", and the `value` will be a string containing the enrollment
  application's provider type.

A `Practitioner` has:
- `identifier`s, as available: NPI, SSN
- `name`

An `Organization` has:
- `identifier`s, as available: NPI, EIN
- `name`

Note that the Employer Identification Number[5] or EIN, referred to in
the PSM as the Federal Employer Identification Number (FEIN), is
represented in the API using the OID assigned to it in the HL7 object
identifier registry[6].

Include unit tests around the transformations and the API itself.

[1] https://www.hl7.org/fhir/task.html
[2] https://www.hl7.org/fhir/practitioner.html
[3] https://www.hl7.org/fhir/organization.html
[4] https://www.hl7.org/fhir/datatypes.html#codeableconcept
[5] https://en.wikipedia.org/wiki/Employer_Identification_Number
[6] https://www.hl7.org/oid/index.cfm?Comp_OID=2.16.840.1.113883.4.4

Issue #641 Create API that returns a list of approved and/or rejected
           providers