Return null if no enrollment with ID is found
Instead of throwing a NullPointerException, check the return value of
the methods `ProviderEnrollmentServiceBean#getTicketDetails(CMSUser,
long)` calls, and handle them returning null appropriately.

It's a broader question of whether returning null is appropriate at all,
but it seems like the least invasive change to fix this broken behavior.

Add tests to verify the new behavior. They are somewhat fragile, as
there are many interactions with the database that need to be mocked
out; as we improve the schema and move more of the joining logic out of
the application and into Hibernate and the database, those tests will
become less fragile (and the code under test will become much smaller).

Found and fixed to support the new FHIR API.

Issue #57 Improve database schema
Issue #641 Create API that returns a list of approved and/or rejected
           providers