Abandon the plan started in commit d3d9c37c
The spreadsheet will remain the master requirements document for now.
The information could be just as easily represented in Org Mode, but
there's no need to make that move unless we find ourselves doing text
editing tasks on requirements often, which so far we haven't, with the
exception of the RTM<->issues mapping work which is almost done now.