Find requirements for future-oriented issues
This takes care of psm-IP-3.7 ("FOR FUTURE FEATURES"), as mentioned in
commit 9ba5431bf.  I had thought we were going to need to make some
new post-1.0 requirements for these issues, but they all fit into
existing requirements pretty well.  It's just that for each of these
issues, the requirement will be satisfied for 1.0 before that
particular issue is resolved.  That's okay; that's probably the case
with many other issues that are now associated with requirements.
When the time comes, we can always split a requirement into pre-1.0
and post-1.0 components if we need to.