Remove colons from individual provider enrollment step pages
Specifically, for a 'Speech Language Pathologist' enrollment.
Includes pages shown when checking yes and no for 'Do you
maintain your own private practice?' Does not include the
'summary' page (to be done separately). Also moves red required
'*' next to the label text on the personal information page,
instead of to the right of the input field, for consistency with
other pages.

Issue #376 Remove colons between form labels and fields