Add a category row to the "SQ" sheet
The new row just says "1.  Software Quality and Maintainability",
which just repeats what the "SQ" sheet is about, but the presence of
this row before any of the actual requirements causes this sheet to
match the pattern of the other sheets, in which a category description
always comes before the requirements in that category.