Remove spurious row from IA sheet
Row 23 in the IA sheet had "How do PSM interface to NPI?" in the
"Source Document" column (column G) but no other fields.  At the very
least this question was in the wrong column, but it also didn't add
anything useful to our understanding of the PSM's requirements, so I
just removed it.