Do "-SA-" to "-SC-" in labels-to-issues.json too
This follows up to commit 1705c5515c, but I wanted this to be a
separate commit because labels-to-issues.json is a historical record
of what we actually did in the issue tracker.  We probably won't need
that record, but if we ever do, the most accurate version of it is the
commit right before this one.