Add "find all" search to Tasks API
Allow retrevial of every enrollment application by GETting
`/cms/fhir/Task/` (with no query parameters).

Because of the current interface and database design, this is an N+1
query: first it gets the list of enrollment applications, and then it
gets the full enrollment data with a separate query for each. This is
not performant! But fixing it would affect several other locations, and
it's low-priority for now; we should fix it when we revise the database
schema (issue #57).

Issue #641 Create API that returns a list of approved and/or rejected
           providers