Add query parameters to Task/enrollment search
Allow searching by provider NPI, type, or name, or by the status of the
enrollment application ("Draft", "Requested", "Rejected", or "Accepted").
Multiple statuses can be specified, which returns applications that have
any status in the list.

Each query parameter can be combined or omitted, leaving the find all
functionality intact.

For example, the following query finds enrollment applications that have
a provider with "p1" in their name (first, last, or organization), that
have an NPI of 1234567893, that have a provider type of Acupuncturist,
and that have been either approved or rejected:

    /cms/fhir/Task/?name=p1&npi=1234567893&providerType=Acupuncturist&status=Accepted,Rejected

Note that the status names are from FHIR, and are not the same as the
status shown in the PSM UI. Searching for an invalid status, or with an
malformed NPI (non-numeric, too short, or too long), will result in an
HTTP 400 Bad Request error, with an appropriate error message.

Issue #641 Create API that returns a list of approved and/or rejected
           providers