Show Facility Credentials in review screen
Provider types that allow uploading of facility credentials - such as
the County Contracted Mental Health Rehab provider type - did not
previously allow those credentials to be shown anywhere. Add a section
to the admin review screen that includes the name, begin date, end date,
and a link to the uploaded attachment.

Note that this only includes "signed contracts" (as the code refers to
them); it is possible that there are other kind of facility credentials
that are not shown by this change.

Issue #261 Facility credentials not shown at review time