Use attachment tag for physician clinic letter
In the facility credentials step in a Physician Clinic enrollment
application, replace the input field with the new attachment tag. This
adds a hidden field with the previous attachment ID; this does not
change the behavior, because the code that handles uploads to this page
does not change the enrollment when it does not get an upload (in
contrast to the other form binders I've seen that handle attachments).

Issue #158 "No file selected" is confusing when a file is uploaded