Use attachment tag for PCPO insurance uploads
Use the new attachment tag in the four insurance fields of the facility
credentials screen of Personal Care Provider Organization enrollment
applications. Wrap this in a floating `<span>` to handle the more-common
way of showing the attachment link in a `<div>` after the `<input>`
rather than without a `<div>` and before the `<input>` element.

Also unlink the `<label>`s from the `<input>`s so that changing the
behavior will be easier later.

Issue #158 "No file selected" is confusing when a file is uploaded