Use more descriptive label on file input
Prevent the browser from showing the "No file selected" message despite
the user having previously uploaded a file by hiding the file `<input>`
element and wrapping it in a `<label>`, as suggested in [1].

The four possible states are:
1. no previous file, no current file (first view) -> prompt for new file
2. no previous file, new current file (after selecting file) -> show new
   filename
3. previous file, no current file (clicking "back") -> show old filename
   with link to download, and set button name to "Replace..."
4. previous file, new current file (clicking "back" and selecting a new
   file) -> show new filename, set button name to "Replace..."

Numbers 1 and 3 are rendered on the server side in the attachment tag,
and the transition to either 2 or 4 is done in JavaScript. Selecting a
new file to upload replaces the link to the existing file in state 3 (or
the empty div in state 1) with the name of the new file.

Thanks to Paul for pairing with me on this!

[1] https://stackoverflow.com/a/17949302/25637

Issue #158 "No file selected" is confusing when a file is uploaded

Co-authored-by: Paul Morris <paul@paulwmorris.com>