Additional email templates
For #341 we would like to send email notifications for various types of application
event.  This commit sets the stage for those notifications by creating a very basic
structure for email templates.

Email content is stored in a static text file, and the subject line and mapping to
those text files is defined in the cms.properties file.  This structure does not
currently dictate a specific templating engine, but the plan is to use the Apache
Velocity Engine.

Co-authored-by: dan9000 <okcomputer9000@gmail.com>