Setup notification service
We would like to have notifications sent through a centralized service, which
will make it easier to improve in future and use in more places. This commit
adds that notification service and registers the bean.

The registration service had its own way of sending notifications, which was used
as the inspiration for the notification service.  Now that notification service has
been placed into its own generic object, we don't need to have notification logic
in the registration service.

This commit also begins to remove some of our reliance of the CMSConfigurator as
a general factory source, relying on injection where possible and moving the
velocity engine to the notification service.

Co-authored-by: dan9000 <okcomputer9000@gmail.com>