List all redundancies introduced by new reqs
We knew that none of our new reqs were redundant with any of the
original non-hidden req set, because that was the set we were
referring to when we decided we needed to create a given new req.
But we didn't know what redundancies there might be between our new
reqs and the 60 old-but-newly-uncovered hidden reqs; now we do.

Rename 'RTM.org' to show its relationship to 'hidden-RTM-rows.org',
and reformat the latter.

Remove the old 'issues-reqs-mapping.org' file, which was out of date
with respect to 'issues-2018-03-31.org' and anyway redundant with the
derived batch input file 'labels-to-issues.json'.

Bring README.md fully up to date with all of the above.