Save copies of notifications
When a notification is sent by the NotificationService we want to save a copy
in a local database.  Some day this may be used for reports or auditing, but
for now we just want to hang onto the messages.

Some emails have passwords in them, and we don't want to store copies
of those passwords.  As a temporary fix the storage method will overwrite
the password before saving the message to the database.

Issue #34 Improve password handling -- this should include a change to this storage
code, as it will be obsolete once passwords are no longer being sent over email.

Issue #341 Notify submitter & drafter via contact email when enrollment status changes