Always run integration tests when invoked
Sometimes, gradle would "know" that "nothing had changed", and refuse to
rerun the integration tests. The integration tests depend on not only
their own source code, but also the rest of the application, and the
application being built and deployed to WildFly. Likewise, sometimes we
want to be able to run the integration tests against an external server.
These considerations prevent us from meaningfully being able to
configure the integration tests' dependencies into gradle.

We were working around that by running the `integration-tests:clean`
target, but the better fix is to tell gradle that it doesn't understand
when the integration tests should be run, and to just run them every
time someone runs the `integration-tests:test` target.