Allow the :aggregate task to run after failed test
We always want to run the `integration-tests:aggregate` task when it is
called, even if the integration tests fail, since it builds the test
reports.

We were previously working around this by invoking gradle twice, once
for the integration tests and once for the test reports; the better
solution is to tell gradle not to block on the integration tests
failing.