Remove outdated business process visualization
The jBPM Eclipse plugin can export a visualization of a business process
notation file, and it seems this was done so, once upon a time. However,
I haven't been able to figure out a way to automate this .bpmn -> .png
process, and the existing visualization is out of sync with the actual
business process. Delete it.

Issue #165 Remove/edit code pointing to old external-sources app