Remove empty background check process step
The last remaining step in the "external systems" part of the enrollment
business process was to run a background check if the applicant was
high-risk. This was never implemented, however; it had only a
placeholder work item handler. Remove the step from the business
process, remove the supporting rule, and remove the attribute from the
process data model.

As the "external systems" set of steps is now empty, also remove the
branching flow in the business process.

Issue #165 Remove/edit code pointing to old external-sources app