Remove legacy NPI lookup verification
One of legacy external data sources was for checking provider NPIs
against the NPPES registry[1]. It turns out it wasn't properly wired up
in the admin review UI: there was no associated Drools rule requiring it
to be checked. Since it wasn't wired up properly, we deleted the data
source, and the replacement we'll eventually create will almost
certainly not work in the same way, remove the old, broken version.

Delete the jBPM work item handler, the API client it depended on, the
XSLT the client used, the relevant bits of the JAXB data model, the MVC
data model attribute, and the Hibernate and PostgreSQL representations.

[1] https://npiregistry.cms.hhs.gov/

Issue #165 Remove/edit code pointing to old external-sources app
Issue #596 Reduce use of JAXB class generation
Issue #761 Save screening results in the database