Add task queueing to RTM
Add an item to the RTM (psm-FR-3.25) to track the fact that state
reviewers will want to be able to pick up new applications and, perhaps,
reassign them to other reviewers in response to certain characteristics.
In essence, the goal of this item is to reduce potential overhead or
duplicate work for state reviewers. The feature for this RTM item is
"Add task assignment and queuing to state reviewer workflow."