Remove non-functional enrollment history step
One of the automatic screenings that seems to have been desired for the
PSM was to find previous enrollments a provider had been approved for.
However, this was never fully implemented; the work item handler was
only a test version, and even that contained an `if` that could never be
true.

Remove the enrollment history step from the business process, delete its
placeholder work item handler, and delete the rules that depended on it
(and so were also non-functional).

Issue #165 Remove/edit code pointing to old external-sources app