Write wrapper method for enrollment notifications: #770
Write a method to add the submitter (the user who submitted the
application) to all notifications about changes to the status of an
enrollment. Wrap modified and submitted notifications with the new
wrapper.  Also, remove an unnecessary variable in AcceptedHandler.

Add first draft messages about application changes to be edited and
updated on a state-by-state basis.

Issue #770 Write notification email templates