Remove legacy license screening step
The last of the legacy automated screening steps is the "Verify License
or Certification" step, which checks each of the submitted licenses
against one of the legacy external data sources. Remove it, and all of
its supporting data and code. Update the license section of the admin
screening UI with something more like the facility licenses, which does
not have a link to automated screening results or a checkbox.

I haven't seen any documentation around how this worked, so this feels
like an appropriate moment to eulogize about its design.

There were a set of rules that mapped each license type to an external
data source, including the URL, the appropriate XSL transforms, and the
data matching approach. The verification step ran after those rules
populated the model with the data the verification step needed, and made
a SOAP call to the specified external data source with the
appropriately-transformed data, transformed the response, and then set a
flag on the license to indicate that it had been verified.

Given how many different license types the PSM knows about, there were a
lot of different external services, a rule for each, and a request XSLT
and a response XSLT for each. There were also several different matching
strategies used by these different licensed types, but not used outside
of this surprisingly complicated feature.

Remove:
- Verify License or Certification step from the business process
  - also remove the converging gateway that disliked having only a
    single path into it, and instead directly connect the LEIE step to
    the step after the "Verification Completed" gateway
- associated jBPM work item handler
- client the work item handler used
  - since this was the last of its kind, also remove its base class. No
    more SOAP!
- rules that mapped license types to external data sources
  - since that was the entirety of the file, also remove the knowledge
    base initialization code
- matchers specified by the rules for parsing the response
- XSL for transforming requests and responses for each license type
- data type for conveying the mapping between license types and data
  sources
- checkbox(es) from admin UI screen
  - as well as the backing MVC data type attribute
- JAXB data type for license verification results
- "verified" attribute from JAXB representation of a license

Issue #165 Remove/edit code pointing to old external-sources app
Issue #596 Reduce use of JAXB class generation
Issue #761 Save screening results in the database