Remove (broken) auto-reject on disclosure step
One of the steps of the business process was to check if the provider
had answered any of the disclosure questions positively, and if so,
automatically reject them. That is not the behavior we want, so delete
the step from the business process, along with the preceding diverging
gateway, and directly connect the previous step with the next step.

Issue #774 Do not reject enrollment on positive provider disclosure