Extract method for repeated error handling code
Several methods had exactly the same code for exactly the same purpose.
Extract a method to reduce that code duplication.

The surrounding code is also duplicated, but because it includes a
`return` statement, it's not quite so easy to extract a method; I also
think keeping the variables named `inputPage` documents intent in a
useful manner, so also leave that alone.

Issue #8 Reduce code duplication