Initial Liquibase migrations
Liquibase is a database migration tool which allows more control over
schema versioning.  It will ultimately replace the use of sql files for
seed and table structure.

Liquibase can be called directly via application code (known as 'automatic
migrations') or separately from the application logic itself (known as
'manual migrations').  This initial implementation leverages the Gradle
infrastructure so that migrations can be made as automatic as desired
just like any other piece of the build and deployment process.

The existing  project is now registered in gradle, which allows clear
migration commands through gradle such as  and .

Gradle now requires a properties file in order to run migration commands.
The database flavor is passed in as a property, in addition to the database
credentials.

Issue #254 Allow database schema changes without losing data