Use persisted LEIE screening data in admin review
Use the newly-persisted LEIE automatic screening data on the admin
review page; it will show the most recent screening result (as there can
now be multiple) as "PASS", "FAIL", "ERROR", or "Not performed" if there
are no related screenings.

As we are accessing the list of automatic screenings outside of the JPA
session, we cannot lazy-load the list; trying to do so raises a
`LazyInitializationException`. Instead, add a new API method that
fetches an enrollment with its screening results.

Issue #761 Save screening results in the database