Remove unused schedule screening endpoint
This was apparently intended to allow an admin to immediately re-run an
automatic screening, but it was never referenced, as confirmed by
`git grep '/schedule'`. Remove it and the (unimplemented) code it
 depended on.

This is part of clearing the way for adding screening persistence and
setting up monthly re-screening.

Issue #355 Improve UI in screening log
Issue #740 Re-screen approved providers monthly
Issue #761 Save screening results in the database