Show persisted automatic screening details
Replace the legacy automatic screening log, which was based on the
JAXB/XML representation of the screening results stored in the jBPM
enrollment process, with a new, LEIE-specific page. As we add more
automatic screening types, we can add more pages, so that each shows the
relevant information.

Issue #355 Improve UI in screening log
Issue #761 Save screening results in the database