Add files for tighter feedback loop in frontend development
Since building and deploying the PSM does not provide a tight
feedback loop for frontend development, this directory and its
files allow for quicker iteration.  Currently that's primarily
for reports and their graphs drawn with d3js.  The files in this
directory are just for development, not meant to ever be merged
into the PSM.