Added roles seed data
This adds support for seed data on the roles table.  Future data can
be added or updated by simply modifying the roles.csv file

Loading of seed data is being handled in a separate changelog, this
is so that we can ensure that seed data is never loaded until after
a table's structure is stable.  This allows us to make sure that
the seed data fits the table's final form (rather than being inserted
into a table that, for instance, doesn't yet have all of its columns)

Imagine a situation where we have change sets in the following order:

A) create the table [id, first_name, last_name]
B) insert data [id, first_name, last_name]
C) add column to table [middle_name]

In this order we can't actually add middle name to the CSV and guarantee it gets inserted when migrations are run, because the middle name column won't exist at the point when the data is inserted.

If we ensure insertion of data always happens at the end then this isn't a problem.

Issue #254 Allow database schema changes without losing data