Show and log exception in common Spring handler
We had two global exception handlers, one for non-admin controllers and
one for admin controllers, which were similar but not the same. Both
referred to the same error view, but populated it with different
variables.

Extract a global exception handler class, using Spring the
`@ControllerAdvice` annotation[1], and update it to log the exception
and the relevant method + URL (but not query parameters, both for
potential privacy reasons and because it is distractingly difficult to
present them well).

Additionally, previously, the error view showed a very generic message.
This is appropriate for a production system, but we're not yet a
production system, and it is frustrating to have such a noninformative
error message.

[1] https://docs.spring.io/spring/docs/4.3.12.RELEASE/spring-framework-reference/html/mvc.html#mvc-ann-controller-advice