Remove unused error page
There was a separate JSP for displaying errors to admin users, but it
doesn't seem to be used anywhere: the admin error handler refers to the
same error page as the non-admin error handler, the string "admin/error"
does not occur in the project, and based on examining all the instances
of the quotes-enclosed string `"error"`, there is no path-construction
going on.

Remove the unused admin version of the error page.