Move reqs code elsewhere; remove CSV files (#799)
The requirements-parsing and requirements-manipulating code is moving
over to the https://github.com/solutionguidance/psm-dashboard project.
We also stop versioning generated CSV exports of requirements here,
because the psm-dashboard now knows how to automatically extract
requirements from RTM.xlsx directly.  (This is good because those CSV
exports were always redundant with RTM.xlsx anyway, and as long as
they were present there was the danger of them becoming out-of-date.)

Most of this change is simple:

Remove a bunch of files from here either because they will live over
in the psm-dashboard (psm-reqs.el, psm_reqs.py, reqs2any, show-reqs),
or because they will no longer exist (the CSV exports).

But one part of this change is complex:

We port over a bunch of edits to requirements/RTM.xlsx that were made
on the old 'rtm-issue-linking' branch but which somehow did not get
merged to 'master', while preserving the one independent edit made to
RTM.xlsx on 'master' after 'rtm-issue-linking' had branched.

Here are the details about that:

A bunch of edits made to RTM.xlsx on the 'rtm-issue-linking' branch
should have come over to 'master' in merge commit 1c097be58b, but for
some reason that merge commit (which squashed a bunch of work from
'rtm-issue-linking') did not include RTM.xlsx, meaning all of the
branch's edits to RTM.xlsx were accidentally stranded.

The most likely explanation is this that we were being careful not
to overwrite commit 548b2c486e on 'master', which changed RTM.xlsx
after 'rtm-issue-linking' had branched (commit 0708e9fc016d is the
last shared commit between 'master' and 'rtm-issue-linking').

However, there were many desirable edits to RTM.xlsx on the branch.
What we should have done was copy the branch's version over to
'master' and then re-done the one edit from 'master' that needed to be
preserved.  That's essentially what we're doing now.

So first, here are all the wonderful edits to 'rtm-issue-linking' that
come over in this commit, in no particular order:

  - commit 1705c5515c

    Change the "Standards & Conditons" sheet's req IDs from "psm-SA-*"
    to "psm-SC-*".  Note that a consequence of this change had already
    made it over to 'master': the requirements/csv/reqs-sc.csv file --
    which although it had "sc" in its name contained req IDs that said
    "SA" until this commit changed them to "SC" -- *did* come over to
    master in commit 1c097be58bc.  This meant that, until now, there
    was a discrepancy between RTM.xls and reqs-sc.csv on 'master'.
    That's really icky, and is evidence that Cecilia and Jason were
    right that versioning the derived CSV files alongside the .xlsx
    was a bad idea all along (so yay that we're going to stop that).

    This edit thus renames the following req IDs:

      psm-SA-1.1  ->  psm-SC-1.1
      psm-SA-2.1  ->  psm-SC-2.1
      psm-SA-3.1  ->  psm-SC-3.1
      psm-SA-4.0  ->  psm-SC-4.1
      psm-SA-4.1  ->  psm-SC-4.1
      psm-SA-4.2  ->  psm-SC-4.2
      psm-SA-4.3  ->  psm-SC-4.3
      psm-SA-4.4  ->  psm-SC-4.4
      psm-SA-4.5  ->  psm-SC-4.5

    (Note that the req ID "PSM-SC-4.0" never existed on master before,
    although the requirement itself did; see commit c077cc0cb below
    for more about this.)

  - commit 9b8eddec6

    Add 'psm-FR-7.15' for providing administrator help documentation

  - commit 18fd980dd

    Add a category row to the "SQ" sheet

    The new row just says "1.  Software Quality and Maintainability",
    which just repeats what the "SQ" sheet is about, but the presence of
    this row before any of the actual requirements causes this sheet to
    match the pattern of the other sheets, in which a category description
    always comes before the requirements in that category.

    (Although this wasn't know at the time the original commit message
    was written, this edit is now especially important because we're
    now using xlsx2csv to export the requirements directly from
    RTM.xlsx.  If this new category row -- row 2, with category
    description in column B -- were not present, the next sheet, which
    is Software Quality ("SQ"), would inherit whatever the last active
    category in the the previous sheet were.  That happens to be
    "5. Technical Service Classification: Security and Privacy" from
    row 23 of the "Access and Delivery" sheet, which is unrelated to
    the SQ sheet and thus would have caused much confusion.)

  - commit c077cc0cb

    Unhide all rows in all tabs in the RTM

    (Although the original commit message did not mention it, this
    commit also gave the req ID "psm-SC-4.0" to the req in row 9 --
    that is, the first req in the "4. S&C: Modularity Standard" group
    -- whose "Requirement ID Number" in column A was blank.)

  - commit 95e3a9f2f

    Update psm-SQ-1.3 to include formats as well as libraries.

    This is to make that req applicable to things like issue #613.

  - commit 14a0c613e

    Label column E as "Rank" in the SQ sheet

    The new "Software Quality" (SQ) sheet should have the same column
    names as it does in every other sheet: "Rank".  However, when we
    created the new sheet (see commit 9ba5431bf and commit 77c8619a8a),
    we forgot to give that column a title.  This commit fixes that.

  - commit b9cbd1ef

    Remove spurious row from IA sheet

    Row 23 in the IA sheet had "How do PSM interface to NPI?" in the
    "Source Document" column (column G) but no other fields.  At the very
    least this question was in the wrong column, but it also didn't add
    anything useful to our understanding of the PSM's requirements, so we
    just removed it.

And here is the one edit preserved from master:

   - commit 548b2c486e

     Add req "psm-FR-3.25", which states "The PSM shall allow state
     reviewers to assign and queue applications" (source: "MT").