Create EFT to FHIR transformer
We are going to provide indication of whether an enrollment
applicant is able to accept EFT payments.  This commit creates
the transformer which will extract that information from a
ProviderProfile's affiliations.

The logic for iterating through affiliations, identifying primary
affiliations, and checking their EFT acceptance is put into the
transformer because it is unclear at this point if this kind of
logic has a home in our entity objects.

Tests for this object are included in this commit.

Issue #841 Add new fields to FHIR API