Invoke EFT transformer in API
We are using FHIR task inputs to pass along data that doesn't yet
have an explicit place in the FHIR standard.  This commit uses the
previously created EnrollmentAcceptsEftToFhir object to include
information about EFT payment capabilities in the API.

It also modifies the tests for the EnrollmentToFhir object both
to validate that the input is included, as well as to make a
previous test less rigid / allow more than one input to be
in the result set.

Issue #846 Add new fields to FHIR API