Fix error to match rule
A facility license needs to have been issued before the "effective date"
for the enrollment application (which makes sense -- you can't become an
authorized facility until you have a license).  The rule worked
correctly, but the error message that appeared when it was broken gave
the opposite message, leading to much consternation for me until I
figured it out.  Fix that so the error text matches the actual rule.