Remove unused legacy indicators
The last bit of legacy information stored was whether or not an entity
was imported from a legacy information source. As the relevant bit of
code that worked with profile importing was deleted in commit
68666be4c698246073b5ee4030a2bffcc76473dd, and the only remaining
references were round-tripping the indicator between the Hibernate and
JAXB data models, they can be deleted, so as to leave us with one fewer
distracting attribute on the Entity model.

Issue #57 Improve database schema
Issue #181 Provider cannot link profile with external auth