Extract class for LEIE lookup from jBPM handler
Move the logic and supporting infrastructure, including the
knowledge of how to construct an LeieAutomaticScreening entity, out of
the jBPM work item handler and into its own, reusuable class. This
should make it easier to screen enrollments in a context outside of the
jBPM enrollment business process - such as an administrator pushing the
re-screening button.

I have intentionally not yet created an interface for this new class to
implement; now that the `ExcludedProvidersScreeningHandler` is
considerably less LEIE-specific, we may want to further genericize it
and reuse it, as it makes fairly clear how to use the JAXB data model.

Issue #740 Re-screen approved providers monthly