Fix NPE on null isEftAccepted
The only place the PSM asks if an organization accepts EFT payments is
for individual providers who own a private practice. An individual
provider attached to a group practice is not asked whether or not that
group practice accepts EFT, and nor are organizational providers[1].

The code to include EFT data in the FHIR API assumes that every
organization either does or does not accept EFT; that approach doesn't
work for organizations for which we have no data.

Fix the null pointer exception resulting from that incomplete
understanding of the world by assuming that an organization which does
not have any data for EFT does not accept EFT. This may not be the
approach we want to stick with in the long run, but it's a reasonable
approach until we make that decision.

[1] https://github.com/SolutionGuidance/psm/issues/411#issuecomment-352091723

Issue #441 Add a "do you accept EFT" checkbox
Issue #846 Add new fields to FHIR API
PR #859 Add EFT Acceptance Indicator to API
Issue #881 FHIR API returns 404 when no value for EFT accepted