Remove unneccessary helper methods
The `renew` and `update` methods added no functionality, and required
extra, unneeded work to fulfill their ignored parameters.

Remove the methods, and replace the calls to them with direct calls to
the `enroll` method they depend on. Also remove the unneccessary profile
lookup.

Although that profile lookup includes a permissions check that the user
can access that profile, the profile we were looking up was the one
attached to the enrollment being submitted. It's been looked up
previously, and we don't need to verify permissions twice.

Issue #740 Re-screen approved providers monthly