Add SpotBugs to our build
Add the static analysis tool SpotBugs to our Java subprojects, so that
we can use it to find errors in our code.

From the website[1]:

    SpotBugs is a program which uses static analysis to look for bugs in
    Java code.

    SpotBugs is the spiritual successor of FindBugs, carrying on from
    the point where it left off with support of its community.

It adds `spotbugs{SourceSetName}` tasks in gradle[2]. In practice, since
we use Groovy for our tests and don't have any other source sets, that
means you can invoke it like so:

    ./gradlew spotbugsMain

Currently, there are many problems in our code that SpotBugs can detect,
so these checks don't pass; I add it now as an aspirational tool that we
can use to find small things to fix in between larger features.

See also the list of bugs it can detect[3].

[1] https://spotbugs.github.io/
[2] https://plugins.gradle.org/plugin/com.github.spotbugs
[3] https://spotbugs.readthedocs.io/en/latest/bugDescriptions.html

Issue #915 Use static analysis to find bugs